Send real UDP packets to Google DNS. Encoding a name correctly in the packet is hard.
Currently using a disciminant to store the length of the string. This is
required by Ada to keep a string in a record if it's not known at
compile time:

```
type Question (Name_Len : Positive) is
    record
        Name : String (1 .. Name_Len);
        ...
    end record;

```

This Looks ok, but in reality there is a problem! The `Name_Len` field
is stored in the record, so when the record is encoded it has extra
bytes on the front. These bytes need to be accounted for if you directly
serialize or deserialize a record. This is not being done correctly
right now.